requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
factory-boy==3.2.1
Faker==14.2.0
idna==3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.12
psycopg2-binary==2.8.6